xrootd: Fix race condition and vector read framing bugs
Merged: trunk@17301
Patch: http://rb.dcache.org/r/4529
Acked-by: Paul Millar
Merge-request: http://rt.dcache.org/Ticket/Display.html?id=7189

+++ modules/dCacheJUnit/org/dcache/pinmanager/PinManagerTests.java
@@ -156,7 +156,7 @@ public class PinManagerTests
         message = processor.messageArrived(message);
 
         assertEquals(0, message.getReturnCode());
-        assertEquals(pin.getPinId(), message.getPinId());
+        assertEquals(pin.getPinId(), (long) message.getPinId());
         assertEquals(pin.getRequestId(), message.getRequestId());
 
         Pin newPin = dao.getPin(pin.getPinId());
@@ -183,7 +183,7 @@ public class PinManagerTests
         message = processor.messageArrived(message);
 
         assertEquals(0, message.getReturnCode());
-        assertEquals(pin.getPinId(), message.getPinId());
+        assertEquals(pin.getPinId(), (long) message.getPinId());
         assertEquals(pin.getRequestId(), message.getRequestId());
 
         Pin newPin = dao.getPin(pin.getPinId());
